+++ lib/beamer/locations/dashboards_location.dart
@@ -2,7 +2,6 @@ import 'package:beamer/beamer.dart';
 import 'package:flutter/material.dart';
 import 'package:lotti/beamer/beamer_delegates.dart';
 import 'package:lotti/pages/create/complete_habit_dialog.dart';
-import 'package:lotti/pages/create/create_measurement_dialog.dart';
 import 'package:lotti/pages/dashboards/dashboard_page.dart';
 import 'package:lotti/pages/dashboards/dashboards_list_page.dart';
 import 'package:lotti/themes/theme.dart';
@@ -15,7 +14,6 @@ class DashboardsLocation extends BeamLocation<BeamState> {
   List<String> get pathPatterns => [
         '/dashboards',
         '/dashboards/:dashboardId',
-        '/dashboards/:dashboardId/measure/:measurableId',
         '/dashboards/:dashboardId/complete_habit/:habitId',
       ];
 
@@ -23,7 +21,6 @@ class DashboardsLocation extends BeamLocation<BeamState> {
   List<BeamPage> buildPages(BuildContext context, BeamState state) {
     bool pathContains(String s) => state.uri.path.contains(s);
     final dashboardId = state.pathParameters['dashboardId'];
-    final measurableId = state.pathParameters['measurableId'];
     final habitId = state.pathParameters['habitId'];
 
     final pages = [
@@ -37,27 +34,6 @@ class DashboardsLocation extends BeamLocation<BeamState> {
         BeamPage(
           key: ValueKey('dashboards-$dashboardId'),
           child: DashboardPage(dashboardId: dashboardId!),
-        ),
-      if ((isUuid(dashboardId)) && measurableId != null && isUuid(measurableId))
-        BeamPage(
-          routeBuilder: (
-            BuildContext context,
-            RouteSettings settings,
-            Widget child,
-          ) {
-            return DialogRoute<void>(
-              context: context,
-              builder: (context) => child,
-              settings: settings,
-              barrierColor: styleConfig().negspace.withOpacity(0.54),
-            );
-          },
-          key: ValueKey('dashboards-measure-$measurableId'),
-          child: MeasurementDialog(measurableId: measurableId),
-          onPopPage: (context, delegate, _, page) {
-            dashboardsBeamerDelegate.beamBack();
-            return false;
-          },
         ),
       if ((isUuid(dashboardId) || pathContains('carousel')) &&
           habitId != null &&